requirements.txt
celery==4.4.0
Django==2.2
django-filter==2.2.0
django-grappelli==2.13.3
django-imagekit==4.0.2
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
djangorestframework-filters==1.0.0.dev0
drf-yasg==1.17.0
fcm-django==0.3.4
Markdown==3.1.1
Pillow==7.0.0
psycopg2==2.8.4
requests==2.22.0
drf-yasg==1.17.0
uWSGI==2.0.18
